Reimplement Nil to be empty list
Nil used to be a ConsC with VOID data. This design is flawed as the
List still contained a single element.

To do this, List::head is now a shared_ptr, that can only be Null if
it is Nil. Since it has a protected default constructor that can only
be invoked by Nil.